+++ src/playgrounds/webview.ts
@@ -19,15 +19,9 @@ export class PlaygroundWebview {
 
   private updateBaseUrl() {
     const owner = this.gist.owner ? this.gist.owner.login : "anonymous";
+    const version = this.gist.history ? this.gist.history[0].version + "/" : "";
 
-    // By default, the gist doesn't include the commit iD for the latest revision,
-    // and so instead of trying to retrieve it from the server, we can simply grab
-    // it from the "url" property.
-    const version = this.gist.url
-      ? this.gist.url.substr(this.gist.url.lastIndexOf("/") + 1)
-      : "";
-
-    this.baseUrl = `https://gist.githack.com/${owner}/${this.gist.id}/raw/${version}/`;
+    this.baseUrl = `https://gist.githack.com/${owner}/${this.gist.id}/raw/${version}`;
   }
 
   constructor(
@@ -58,19 +52,24 @@ export class PlaygroundWebview {
             url: value.url,
             method: value.method,
             data: value.body,
-            headers: JSON.parse(value.headers || {})
+            headers: JSON.parse(value.headers || {}),
+            responseType: "text",
+            transformResponse: (data) => {
+              return data;
+            }
           });
 
           webview.postMessage({
             command: "httpResponse",
             value: {
               id: value.id,
-              body: JSON.stringify(response.data),
+              body: response.data,
               status: response.status,
               statusText: response.statusText,
               headers: JSON.stringify(response.headers || {})
             }
           });
+
           break;
       }
     });